First working version of all code.
ime_explore_parameter_space.pl is a wrapper script around ime_trainer.pl to help explore parameters.

imeter.pl is a renamed version of what was called 'the_imeter.pl' and now includes the latest
parameters that we will use to train the IMEter in other species. This parameter file is included
along with the test of wild-type introns that we used to calibrate the IMEter against known
expression scores.

We also include the set of A. thaliana introns, exons, and UTRs produced by proc_phytozome.pl